# Don't call escapeString() on time values