@inconshreveable 2016-06-21
i like it, but we should probably add a comment that it's for performance reasons