@tg 2016-06-21
I guess an easier (and cleaner) solution would be to simply add a `case` to `select` in `formatLogfmtValue` which, in case of `time.Time`, just returns the value retrieved from `formatShared`:

```
case time.Time:
	// No need for escaping, just return raw string value
	return value
```