@kevinburke 2016-06-22
> I guess an easier (and cleaner) solution would be to simply add a case to select in formatLogfmtValue which, in case of time.Time, just returns the value retrieved from formatShared:

I think the problem is `formatShared` takes a `time.Time` as input and returns a string, so there's not a good way to check whether the value returned from `formatShared` was originally a time.Time or whatever.