@kevinburke 2016-06-22
We could save it in a different value and then compare against it, or have formatShared return a time.Time, but either implies more work