@kevinburke 2016-06-22
98d12c2 updates the comment