@tg 2016-06-22
@kevinburke you're right, I should have read the code more carefully.